moved the active  collection card component to a shared folder and made this component more versatile by adding mor props, also changed the import
