+++ src/modules/mainPageCollection/components/ActiveCollectionList/ActiveCollectionList.jsx
@@ -1,16 +1,18 @@
 import React from 'react';
 import clsx from 'clsx';
-import { ActiveCollectionsCard } from '@/modules/mainPageCollection';
+import { ActiveCollectionsCard } from '@/shared/components';
 import s from './ActiveCollectionList.module.scss';
 
-const ActiveCollectionList = ({ visibleItems, collection, className }) => {
-  const collections = collection.collections;
-
+const ActiveCollectionList = ({ visibleItems, allCollection, className }) => {
   return (
     <ul className={clsx(s.collectionList, className)}>
-      {collections.slice(0, visibleItems).map((collection) => (
+      {allCollection?.collections?.slice(0, visibleItems).map((collection) => (
         <li key={collection._id}>
-          <ActiveCollectionsCard collection={collection} />
+          <ActiveCollectionsCard
+            collection={collection}
+            buttonDetails={allCollection.button_details}
+            buttonDonas={allCollection.button_donas}
+          />
         </li>
       ))}
     </ul>
