+++ src/modules/mainPageCollection/components/MainPageCollection/MainPageCollection.jsx
@@ -3,11 +3,15 @@
 import { useEffect, useState } from 'react';
 import { useRouter } from 'next/navigation';
 import { SwiperSlide } from 'swiper/react';
-import { Button, Container, DotsSwiper } from '@/shared/components';
+import {
+  ActiveCollectionsCard,
+  Button,
+  Container,
+  DotsSwiper,
+} from '@/shared/components';
 import { ROUTES } from '@/shared/constants';
 import { ActiveCollectionList } from '@/modules/mainPageCollection';
 import collections from '../../data/section-content.json';
-import { ActiveCollectionsCard } from '@/modules/mainPageCollection';
 import s from './MainPageCollection.module.scss';
 
 const MainPageCollection = () => {
@@ -56,7 +60,7 @@ const MainPageCollection = () => {
             ) : (
               <ActiveCollectionList
                 visibleItems={visibleItems}
-                collection={collections}
+                allCollection={collections}
               />
             )}
           </div>
