+++ src/modules/mainPageCollection/index.js
@@ -1,4 +1,3 @@
 export { default as MainPageCollection } from './components/MainPageCollection/MainPageCollection';
-export { default as ActiveCollectionsCard } from './components/ActiveCollectionsCard/ActiveCollectionsCard';
 export { default as ActiveCollectionList } from './components/ActiveCollectionList/ActiveCollectionList';
 export { default as CollectionProgress } from './components/CollectionProgress/CollectionProgress';
