+++ src/shared/components/ActiveCollectionsCard/ActiveCollectionsCard.module.scss
@@ -1,5 +1,3 @@
-@import '../../../../shared/styles/mixins';
-
 .ContentContainer {
   padding: 16px 16px 24px 16px;
   border-radius: 32px;
