+++ src/shared/components/index.js
@@ -12,4 +12,5 @@ export { default as PublicPrivateForm } from './PublicPrivateForm/PublicPrivateF
 export { default as DotsSwiper } from './DotsSwiper/DotsSwiper.jsx';
 export { default as CustomPagination } from './DotsSwiper/CustomPagination/CustomPagination';
 export { default as ProgressBar } from './ProgressBar/ProgressBar';
-export {default as BreadCrumbs} from './Breadcrumbs/Breadcrumbs.jsx'
+export { default as BreadCrumbs } from './Breadcrumbs/Breadcrumbs.jsx';
+export { default as ActiveCollectionsCard } from './ActiveCollectionsCard/ActiveCollectionsCard';
